move term to safekeeper_api
