+++ libs/safekeeper_api/src/lib.rs
@@ -5,6 +5,12 @@ use const_format::formatcp;
 /// Public API types
 pub mod models;
 
+pub mod state;
+
+/// Consensus logical timestamp. Note: it is a part of sk control file.
+pub type Term = u64;
+pub const INVALID_TERM: Term = 0;
+
 pub const DEFAULT_PG_LISTEN_PORT: u16 = 5454;
 pub const DEFAULT_PG_LISTEN_ADDR: &str = formatcp!("127.0.0.1:{DEFAULT_PG_LISTEN_PORT}");
