+++ libs/safekeeper_api/src/state.rs
@@ -1,0 +1,1 @@
+
